Add page listing all annotations
